Create relationships between ORM classes (#203)
Update the ORM objects to include relationships between each table.

This also includes adding a test that reads a person and all of their relationships. Consequently, this PR also includes a dummy _convert_ingest_info() to enable tests that verify that relationships are correctly matched. This logic will need to be completely updated with the data converter logic (ie. ingest_info -> database schema) which is currently being written.

PAIR = @terinpw