Update data converter to use ingest_info.proto (#303)
This PR updates the data converter to convert `ingest_info.proto` -> `persistence layer entities` instead of `ingest_info.py` -> `schema ORM object`.

As explained in #302, we should consider breaking apart this converter to further isolate code (this file is starting to contain a lot of logic). This has the additional benefit of allowing us to write more detailed tests against each compontent (eg. when booking has `age` and `DOB`). We should write those tests in the future. Note that these changes are not included now in the interest of testing an end-to-end prod run.