Update travis to use docker (#1042)
Note: We acknowledge that the time needed to download the cached docker image is nearly the same as the time needed to download the dependencies directly. Through testing we've determined that using the cached image is slightly faster than rebuilding the image entirely (~180 sec vs ~215 sec). If this changes, we should stop using the Travis cache.

If we wanted to further reduce the time to build our docker image, we should consider moving off the base Ubuntu image.